requirements.txt
streamlit
langchain
langchain-community
faiss-cpu
sentence-transformers
langchain-google-genai
google-cloud-firestore